move sshclient.py to helper
